Refactor exceptions
- Moved `Phly\Mustache\Exception` interface to
  `Phly\Mustache\Exception\ExceptionInterface`.
- Updated all extensions to reflect the above.
- Base all exceptions on more specific SPL exceptions instead of global
  exception class.
